change int to long int

+++ src/CUDACalculator.cu
@@ -10,14 +10,14 @@ using namespace cv;
 static const int blockSize = 1024;
 static const int gridSize = 12; //this number is hardware-dependent; usually #SM*2 is a good number.
 
-__global__ void sumCommMultiBlock(const int *gArr, int arraySize, int *gOut) {
+__global__ void sumCommMultiBlock(const long int *gArr, int arraySize, long int *gOut) {
     int thIdx = threadIdx.x;
     int gthIdx = thIdx + blockIdx.x*blockSize;
     const int gridSize = blockSize*gridDim.x;
-    int sum = 0;
+    long int sum = 0;
     for (int i = gthIdx; i < arraySize; i += gridSize)
         sum += gArr[i];
-    __shared__ int shArr[blockSize];
+    __shared__ long int shArr[blockSize];
     shArr[thIdx] = sum;
     __syncthreads();
     for (int size = blockSize/2; size>0; size/=2) { //uniform
@@ -29,7 +29,7 @@ __global__ void sumCommMultiBlock(const int *gArr, int arraySize, int *gOut) {
         gOut[blockIdx.x] = shArr[0];
 }
 
-int sumArray(cv::Mat roi_domain, int arraySize) {
+long int sumArray(cv::Mat roi_domain, int arraySize) {
 //    int* roi_in;
 //    cudaMalloc(&roi_in, arraySize);
 //    cudaMemcpy(roi_in, roi_domain.data, arraySize, cudaMemcpyHostToDevice);
@@ -37,19 +37,19 @@ int sumArray(cv::Mat roi_domain, int arraySize) {
     cudaDeviceSetCacheConfig(cudaFuncCachePreferL1);
     cudaDeviceSetSharedMemConfig(cudaSharedMemBankSizeEightByte);
     // transfer data from host to device
-    int* img_in;
-    cudaMalloc(&img_in, arraySize * sizeof(int));
-    cudaMemcpy(img_in, roi_domain.data, arraySize * sizeof(int), cudaMemcpyHostToDevice);
+    long int* img_in;
+    cudaMalloc(&img_in, arraySize * sizeof(long int));
+    cudaMemcpy(img_in, roi_domain.data, arraySize * sizeof(long int), cudaMemcpyHostToDevice);
 
-    int roi_sum;
-    int* img_out;
-    cudaMalloc((void**)&img_out, sizeof(int)*gridSize);
+    long int roi_sum;
+    long int* img_out;
+    cudaMalloc((void**)&img_out, sizeof(long int)*gridSize);
     // call the sum function
     sumCommMultiBlock<<<gridSize, blockSize>>>(img_in, arraySize, img_out);
     sumCommMultiBlock<<<1, blockSize>>>(img_out, gridSize, img_out);
     cudaDeviceSynchronize();
     // transfer data from device to host
-    cudaMemcpy(&roi_sum, img_out, sizeof(int), cudaMemcpyDeviceToHost);
+    cudaMemcpy(&roi_sum, img_out, sizeof(long int), cudaMemcpyDeviceToHost);
     cudaFree(img_in);
     cudaFree(img_out);
     return roi_sum;
